fix wrong data type for automation trigger

+++ esphome/components/microphone/__init__.py
@@ -54,10 +54,9 @@ IsMutedCondition = microphone_ns.class_("IsMutedCondition", automation.Condition
 async def setup_microphone_core_(var, config):
     for conf in config.get(CONF_ON_DATA, []):
         trigger = cg.new_Pvariable(conf[CONF_TRIGGER_ID], var)
-        # Future PR will change the vector type to uint8
         await automation.build_automation(
             trigger,
-            [(cg.std_vector.template(cg.int16).operator("ref").operator("const"), "x")],
+            [(cg.std_vector.template(cg.uint8).operator("ref").operator("const"), "x")],
             conf,
         )
